migration mistake

+++ db/schema.rb
@@ -20,7 +20,7 @@ ActiveRecord::Schema.define(version: 2022_04_19_075726) do
     t.boolean "correct", default: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "question_id", null: false
+    t.bigint "question_id", null: false
     t.index ["question_id"], name: "index_answers_on_question_id"
   end
 
@@ -32,8 +32,8 @@ ActiveRecord::Schema.define(version: 2022_04_19_075726) do
 
   create_table "gists", force: :cascade do |t|
     t.string "url", null: false
-    t.integer "question_id", null: false
-    t.integer "author_id", null: false
+    t.bigint "question_id", null: false
+    t.bigint "author_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["author_id"], name: "index_gists_on_author_id"
@@ -44,16 +44,16 @@ ActiveRecord::Schema.define(version: 2022_04_19_075726) do
     t.string "body", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "test_id", null: false
+    t.bigint "test_id", null: false
     t.index ["test_id"], name: "index_questions_on_test_id"
   end
 
   create_table "test_completions", force: :cascade do |t|
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "user_id", null: false
-    t.integer "test_id", null: false
-    t.integer "current_question_id"
+    t.bigint "user_id", null: false
+    t.bigint "test_id", null: false
+    t.bigint "current_question_id"
     t.integer "correct_questions", default: 0
     t.index ["current_question_id"], name: "index_test_completions_on_current_question_id"
     t.index ["test_id"], name: "index_test_completions_on_test_id"
@@ -65,8 +65,8 @@ ActiveRecord::Schema.define(version: 2022_04_19_075726) do
     t.integer "level", default: 0, null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "author_id", null: false
-    t.integer "category_id", null: false
+    t.bigint "author_id", null: false
+    t.bigint "category_id", null: false
     t.index ["author_id"], name: "index_tests_on_author_id"
     t.index ["category_id"], name: "index_tests_on_category_id"
     t.index ["title", "level"], name: "index_tests_on_title_and_level", unique: true
